+++ codegen/gql_code_builder/CHANGELOG.md
@@ -1,3 +1,17 @@
+## 0.14.0
+
+- Fixes Interface Fragment Inheritance (#484)
+- Major refactor (#485)
+
+## 0.13.1
+
+- support analyzer 7.0.0
+
+## 0.13.0
+
+- allow importing an externally generated schema
+
+
 ## 0.12.0
 
 - Moving built-in Serializers to gql_code_builder_serializers package, so